process all the actions in bulk
This prevents a situation where an action matches a saga that emits an
action, which is picked up in the second emit, which then removes and
re-adds the listener, when the first emit comes back if it also matches
the saga then the listener reference does not exist, but the take re-adds
which results in the take being duplicated and firing twice for every
action.

Instead we prevent this scenario and ignore the sub action emit.